# backend/routes/upload.py
# Copyright 2025 Alun King
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

from fastapi import APIRouter, UploadFile, File, Depends, HTTPException, Query
from sqlalchemy.orm import Session
from sqlalchemy import and_, exists
from backend.config import settings
from backend.models import RawFile, RawFileOut, RawFileType, GroupMember, User, Group, Meeting
from werkzeug.utils import secure_filename
from backend.db_dependency import get_db
from backend.jobs.service import submit_job
from fastapi.responses import FileResponse
import logging
from backend.auth import (
    EmailPrincipal,
    get_current_user_id,
    get_group_role,
    is_email_workflow_group_member,
    is_group_member,
)
import uuid
import os
import re
from pathlib import Path

logger = logging.getLogger(__name__)

import shutil
import pysubs2
#from backend.processing import transcribe

router = APIRouter()

# Allowed file extensions
ALLOWED_AUDIO_EXTENSIONS = {'.wav', '.mp3'}
ALLOWED_TRANSCRIPT_EXTENSIONS = {'.vtt', '.srt'}
ALL_ALLOWED_EXTENSIONS = ALLOWED_AUDIO_EXTENSIONS | ALLOWED_TRANSCRIPT_EXTENSIONS
FILENAME_RE = re.compile(r"^[\w.\-]+$")  # Allows: a-zA-Z0-9 _ . -

def validate_filename(filename: str) -> str:
    if not FILENAME_RE.fullmatch(filename):
        raise HTTPException(status_code=400, detail="Invalid filename")
    return filename


@router.post("/groups/{group_id}/meetings/{meeting_id}/upload/", tags=["meetings"], response_model=RawFileOut)
async def upload_file(
        group_id:int,
        meeting_id:int,
        file: UploadFile = File(...),
        db: Session = Depends(get_db),
        principal: EmailPrincipal = Depends(is_email_workflow_group_member)
    ):
    # Extract extension and validate
    ext = os.path.splitext(file.filename)[1].lower()
    if ext not in ALL_ALLOWED_EXTENSIONS:
        raise HTTPException(
            status_code=400,
            detail=f"Unsupported file type: '{ext}'. Allowed types: {', '.join(sorted(ALL_ALLOWED_EXTENSIONS))}"
        )

    if principal.scoped_group_member_id is not None and ext in ALLOWED_AUDIO_EXTENSIONS:
        raise HTTPException(status_code=403, detail="Group members may only upload transcripts")

    meeting = db.query(Meeting).filter(
        Meeting.id == meeting_id,
        Meeting.group_id == group_id,
    ).first()
    if not meeting:
        raise HTTPException(status_code=404, detail="Meeting not found")
    
    # if it's not an audio file, then it's a provided transcript
    file_type=RawFileType.TRANSCRIPT_PROVIDED
    if ext in ALLOWED_AUDIO_EXTENSIONS:
        file_type=RawFileType.AUDIO
    
    human_filename = secure_filename(file.filename)
    safe_filename = f"{uuid.uuid4().hex}_{human_filename}"
    
    # Build target directory path: uploads/group_id/meeting_id
    target_dir = settings.UPLOAD_DIR / str(group_id) / str(meeting_id)
    target_dir.mkdir(parents=True, exist_ok=True)

    file_path = target_dir / safe_filename

    with open(file_path, "wb") as buffer:
        shutil.copyfileobj(file.file, buffer)

    if ext == ".srt":
        converted_filename = f"{Path(safe_filename).stem}.vtt"
        converted_path = target_dir / converted_filename
        try:
            subtitles = pysubs2.load(str(file_path), encoding="utf-8")
            subtitles.save(str(converted_path), format_="vtt")
        except Exception as exc:
            file_path.unlink(missing_ok=True)
            converted_path.unlink(missing_ok=True)
            raise HTTPException(status_code=400, detail="Invalid SRT transcript") from exc

        file_path.unlink()
        safe_filename = converted_filename
        file_path = converted_path

    # add file upload to the database. Leave processed as null as it has not yet been looked at.
    
    raw_file = RawFile(
        file_name=safe_filename,
        human_name=human_filename,
        description=None,
        meeting_id = meeting_id,
        type=file_type,
    )
    db.add(raw_file)
    db.commit()
    db.refresh(raw_file)

    if file_path.suffix.lower() == ".vtt" and file_type == RawFileType.TRANSCRIPT_PROVIDED:
        group = db.query(Group).get(group_id)
        meeting = db.query(Meeting).get(meeting_id)
        if group is not None and meeting is not None:
            job = submit_job(
                db, "transcript_processing",
                {
                    "group_id": group_id, "meeting_id": meeting_id, "transcript_path": str(file_path),
                    "group_name": group.name, "meeting_date": meeting.date.date().isoformat(),
                },
                user_id=principal.user_id, group_member_id=principal.scoped_group_member_id,
                group_id=group_id, meeting_id=meeting_id,
            )
            raw_file.processing_job_id = job.id

    return raw_file

    # Trigger processing (placeholder)
    # result = process_audio(file_path)
@router.get("/files/media/{group_id}/{meeting_id}/{filename}")
def serve_media(
        group_id:int,
        meeting_id:int,
        filename:str = Depends(validate_filename),
        user_id: int = Depends(get_current_user_id),
        db: Session = Depends(get_db)
    ):

    if not user_can_access_group(db, user_id, group_id):
        raise HTTPException(status_code=403, detail="Forbidden")
    file_path = os.path.join(settings.UPLOAD_DIR, str(group_id), str(meeting_id), filename)

    if not os.path.exists(file_path):
        raise HTTPException(status_code=404, detail="File not found: "+file_path)
    
    response = FileResponse(file_path)
    return response
    
@router.get("/files/embedding/{filename}")
def serve_file(
        filename: str = Depends(validate_filename), 
        user_id: int = Depends(get_current_user_id),
        db: Session = Depends(get_db)
    ):

    if not user_can_access_embedding(db, user_id, filename):
        raise HTTPException(status_code=403, detail="Forbidden")
    file_path = os.path.join(settings.EMBEDDING_DIR, filename)

    if not os.path.exists(file_path):
        raise HTTPException(status_code=404, detail="File not found: "+file_path)
    
    response = FileResponse(file_path)
    return response

def user_can_access_embedding(db: Session, user_id: int, filename: str) -> bool:
    # Find the GroupMember(s) that own this file
    member = db.query(GroupMember).filter(GroupMember.embedding_audio_path == filename).first()
    if not member:
        return False

    # Get groups of the member
    member_group_ids = [group.id for group in member.groups]

    # Get groups of the user
    user_group_ids = (
        db.query(User)
        .filter(User.id == user_id)
        .join(User.groups)
        .with_entities(Group.id)
        .all()
    )
    user_group_ids = [g[0] for g in user_group_ids]

    # Check if any group overlaps
    return any(gid in user_group_ids for gid in member_group_ids)

def user_can_access_group(db: Session, user_id: int, group_id: int) -> bool:
    return db.query(
        exists().where(
            Group.id == group_id,
            Group.users.any(User.id == user_id)
        )
    ).scalar()