# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_file_groups__group_id__meetings__meeting_id__upload__post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_file_groups__group_id__meetings__meeting_id__upload__post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobAccepted": {
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "state": {
            "default": "queued",
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "job_id"
        ],
        "title": "JobAccepted",
        "type": "object"
      },
...
      "RawFileOut": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "file_name": {
            "title": "File Name",
            "type": "string"
          },
          "human_name": {
            "title": "Human Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "processed_date": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processed Date"
          },
          "processing_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing Job Id"
          },
          "type": {
            "$ref": "#/components/schemas/RawFileType"
          }
        },
        "required": [
          "id",
          "human_name",
          "file_name",
          "description",
          "processed_date",
          "type"
        ],
        "title": "RawFileOut",
        "type": "object"
      },
      "RawFileType": {
        "enum": [
          "audio",
          "transcript_provided",
          "transcript_generated"
        ],
        "title": "RawFileType",
        "type": "string"
      },
      "ReportAnswerRequest": {
        "properties": {
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/ReportEvidenceItem"
            },
            "title": "Evidence",
            "type": "array"
          },
          "period_end": {
            "format": "date",
            "title": "Period End",
            "type": "string"
          },
          "period_start": {
            "format": "date",
            "title": "Period Start",
            "type": "string"
          },
          "question": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question",
          "period_start",
          "period_end",
          "evidence"
        ],
        "title": "ReportAnswerRequest",
        "type": "object"
      },
      "ReportAnswerResponse": {
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          }
        },
        "required": [
          "answer",
          "model"
        ],
        "title": "ReportAnswerResponse",
        "type": "object"
      },
      "ReportEvidenceItem": {
        "properties": {
          "citation": {
            "title": "Citation",
            "type": "string"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "event_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Date"
          },
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "source": {
            "description": "meetings, github or trello",
            "title": "Source",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "source",
          "evidence_id",
          "title",
          "content",
          "citation"
        ],
        "title": "ReportEvidenceItem",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WeeklyReportRequest": {
        "properties": {
          "period_end": {
            "description": "Exclusive: meetings dated before this date.",
            "format": "date",
            "title": "Period End",
            "type": "string"
          },
          "period_start": {
            "format": "date",
            "title": "Period Start",
            "type": "string"
          }
        },
        "required": [
          "period_start",
          "period_end"
        ],
        "title": "WeeklyReportRequest",
        "type": "object"
      },
      "WeeklyReportResponse": {
        "properties": {
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/ReportEvidenceItem"
            },
            "title": "Evidence",
            "type": "array"
          },
          "group_id": {
            "title": "Group Id",
            "type": "integer"
          },
          "group_name": {
            "title": "Group Name",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "period_end": {
            "format": "date",
            "title": "Period End",
            "type": "string"
          },
          "period_start": {
            "format": "date",
            "title": "Period Start",
            "type": "string"
          },
          "report_text": {
            "title": "Report Text",
            "type": "string"
          },
          "unavailable": {
            "description": "Linked sources that could not be read for this report.",
            "items": {
              "type": "string"
            },
            "title": "Unavailable",
            "type": "array"
          }
        },
        "required": [
          "group_id",
          "group_name",
          "period_start",
          "period_end",
          "report_text",
          "model",
          "evidence"
        ],
        "title": "WeeklyReportResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/users/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/files/embedding/{filename}": {
      "get": {
        "operationId": "serve_file_files_embedding__filename__get",
        "parameters": [
          {
            "in": "path",
            "name": "filename",
            "required": true,
            "schema": {
              "title": "Filename",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Serve File"
      }
    },
    "/files/media/{group_id}/{meeting_id}/{filename}": {
      "get": {
        "operationId": "serve_media_files_media__group_id___meeting_id___filename__get",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "title": "Meeting Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "filename",
            "required": true,
            "schema": {
              "title": "Filename",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Serve Media"
      }
    },
...
    "/groups/{group_id}/meetings/{meeting_id}/upload/": {
      "post": {
        "operationId": "upload_file_groups__group_id__meetings__meeting_id__upload__post",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "title": "Meeting Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_groups__group_id__meetings__meeting_id__upload__post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RawFileOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Upload File",
        "tags": [
          "meetings"
        ]
      }
    },
...
    "/groups/{group_id}/reports/answer": {
      "post": {
        "description": "Answer a follow-up question using only the evidence from a previously composed report.",
        "operationId": "answer_question_groups__group_id__reports_answer_post",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportAnswerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportAnswerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Answer Question",
        "tags": [
          "reports"
        ]
      }
    },
    "/groups/{group_id}/reports/weekly": {
      "post": {
        "description": "Compose a cited report for meetings dated in [period_start, period_end) plus GitHub and\nTrello activity. Stateless: nothing is stored or emailed; the caller owns delivery.",
        "operationId": "weekly_report_groups__group_id__reports_weekly_post",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "integer"
            }
          },
          {
            "description": "Run as a background job and return 202 with a job id to poll at GET /jobs/{id}.",
            "in": "query",
            "name": "async",
            "required": false,
            "schema": {
              "default": false,
              "description": "Run as a background job and return 202 with a job id to poll at GET /jobs/{id}.",
              "title": "Async",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeeklyReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            },
            "description": "Accepted"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Weekly Report",
        "tags": [
          "reports"
        ]
      }
    },
...
  }
}